requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-pptx
python-docx
PyMuPDF
python-multipart
tqdm
supabase
